Add a lib
