+++ rs/backend/Cargo.toml
@@ -5,9 +5,6 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
 base64 = "0.13.0"
 candid = "0.8.1"
