+++ rs/backend/src/bin/nns-dapp-check-args.rs
@@ -1,0 +1,6 @@
+// use nns_dapp::TestType;
+
+
+fn main() {
+    println!("Hello world");
+}
